feat: migrate code to files according to the MVC model
